==> 0_language_analysis.ipynb <==
import pandas as pd

# Languages

langs = pd.read_csv("data/MUSIC4ALL/raw/id_lang.csv", sep='\t')
langs.head()

langs.info()

langs.describe()

print(sorted(langs.lang.unique().tolist()))

langs.lang.value_counts()
