==> 0_metadata_analysis.ipynb <==
import pandas as pd

# Metadata

metadata = pd.read_csv("data/MUSIC4ALL/raw/id_metadata.csv", sep='\t')
metadata.head()

# Looking for duplicated ids
metadata[metadata.duplicated('spotify_id')].head()

# Duplicated songs needs futher inspection
metadata[metadata.spotify_id=='0iHZnbbFHDO9MzHWlJI0Hb']

metadata.info()

metadata.describe()
