==> 0_music4all_information_analysis.ipynb <==
import pandas as pd

# Information

information = pd.read_csv("data/MUSIC4ALL/raw/id_information.csv", sep='\t')
information.head()

information.info()

information.describe()

information.artist.value_counts()

information.album_name.value_counts()

information.album_name.nunique()
