==> 1_lyrics_gathering.ipynb <==
import pandas as pd
from tqdm import tqdm

import glob
from os.path import basename

lyrics = pd.DataFrame(columns=['id', 'lyrics'])

# Reading all txt files with lyrics and adding to a single dataframe

for txt_path in tqdm(glob.glob("data/MUSIC4ALL/lyrics/*.txt")):
    with open(txt_path, 'r') as f:
        content = f.read()
        id_nb = basename(txt_path).split('.')[0]
        lyrics.loc[len(lyrics)] = [id_nb, content]

lyrics.head()

lyrics.info()

lyrics.to_csv("data/MUSIC4ALL/processed/id_lyrics.csv", index=True)
