==> 2_instrumental_filtering.ipynb <==
import pandas as pd

# Raw Data

langs = pd.read_csv("data/MUSIC4ALL/raw/id_lang.csv", sep='\t')
information = pd.read_csv("data/MUSIC4ALL/raw/id_information.csv", sep='\t')
genres = pd.read_csv("data/MUSIC4ALL/raw/id_genres.csv", sep='\t')
metadata = pd.read_csv("data/MUSIC4ALL/raw/id_metadata.csv", sep='\t')

metadata_list = [langs, information, genres, metadata]

# Processed Data

lyrics = pd.read_csv("data/MUSIC4ALL/processed/id_lyrics.csv", index_col=0)

# Merge Data

merged_data = lyrics.copy()

for df in metadata_list:
    merged_data = merged_data.merge(df, on='id')

merged_data.head()

merged_data.info()

# Filtering Out Instrumental Songs

filtered_data_by_lyrics = merged_data[merged_data.lyrics.str.lower().str.strip()!='instrumental']

filtered_data_by_lang = merged_data[merged_data.lang.str.lower().str.strip()!='intrumental']

len(filtered_data_by_lyrics)==len(filtered_data_by_lang)

len(filtered_data_by_lyrics)

len(filtered_data_by_lang)

# Investigate difference between sets of instrumental songs

set(filtered_data_by_lang.index) - set(filtered_data_by_lyrics.index)

set(filtered_data_by_lyrics.index) - set(filtered_data_by_lang.index)

filtered_data_by_lang[filtered_data_by_lang.index.isin([5280, 101119])]

Obs:
- The smallest set shall be chosen
- What's and id language?

# Investigate id lang examples

data_id_lang = merged_data[merged_data.lang.str.lower().str.strip()=='id']

data_id_lang.head()

# Writing the smallest set
filtered_data_by_lyrics.to_csv('data/MUSIC4ALL/processed/filtered_music4all_v1.csv', index=True)
