==> 3_lang_dectection.ipynb <==
import pandas as pd

import langdetect as ld
import spacy
import spacy_fastlang

from tqdm.notebook import tqdm

#from spacy_langdetect import LanguageDetector
#from spacy.language import Language
#from ast import literal_eval

tqdm.pandas()

# Language Detection

data = pd.read_csv('data/MUSIC4ALL/processed/filtered_music4all_v1.csv', index_col=0)

## Google Langdetect

data["google_lang"] = data.lyrics.progress_apply(ld.detect)

data.head()

## FastText Langdetect

## Uncomment this line to download the model, then comment it again
#!python -m spacy download pt_core_news_md

fasttext_nlp = spacy.load('pt_core_news_md')
fasttext_nlp.add_pipe("language_detector")

def get_fasttext_lang(text):
    return fasttext_nlp(text)._.language

data["fasttext_lang"] = data.lyrics.progress_apply(get_fasttext_lang)

data.head()

data.to_csv("data/MUSIC4ALL/processed/filtered_music4all_v2.csv", index=True)
