==> 3_lang_filtering.ipynb <==
import pandas as pd

# Filtering songs by language

data = pd.read_csv("data/MUSIC4ALL/processed/filtered_music4all_v2.csv", index_col=0)

data.info()

data.describe()

# Investigate portuguese data

### Songs with all langs as PT

three_way_pt_mask = (data.google_lang=='pt')&(data.fasttext_lang=='pt')&(data.lang=='pt')
three_way_pt = data[three_way_pt_mask]
len(three_way_pt)

### Songs with fast text and original attribute as PT

lang_fast_pt_mask = (data.google_lang!='pt')&(data.fasttext_lang=='pt')&(data.lang=='pt')
lang_fast_pt = data[lang_fast_pt_mask]
len(lang_fast_pt)

### Songs with langdetect and original attribute as PT

lang_google_pt_mask = (data.google_lang=='pt')&(data.fasttext_lang!='pt')&(data.lang=='pt')
lang_google_pt = data[lang_google_pt_mask]
len(lang_google_pt)

lang_google_pt[['lyrics', 'artist', 'song', 'lang', 'google_lang', 'fasttext_lang']]

# Manual inspection on lyrics
SELECTED_LANG_GOOGLE = [45270, 68930, 91661]

# Songs to keep
selected_lang_google_pt = lang_google_pt[lang_google_pt.index.isin(SELECTED_LANG_GOOGLE)]
len(selected_lang_google_pt)

### Songs with fast text and langdetect as PT

fasttext_google_pt_mask = (data.google_lang=='pt')&(data.fasttext_lang=='pt')&(data.lang!='pt')
fasttext_google_pt = data[fasttext_google_pt_mask]
len(fasttext_google_pt)

### Songs with only fast text as PT

fasttext_pt_mask = (data.google_lang!='pt')&(data.fasttext_lang=='pt')&(data.lang!='pt')
fasttext_pt = data[fasttext_pt_mask]
len(fasttext_pt)

fasttext_pt[['lyrics', 'artist', 'song', 'lang', 'google_lang', 'fasttext_lang']]

# Songs to remove
SELECTED_FASTTEXT = [26733, 35071, 61970, 65163, 75828, 91891]

selected_fasttext_pt = fasttext_pt[~fasttext_pt.index.isin(SELECTED_FASTTEXT)]
len(selected_fasttext_pt)

### Songs with only langdetect as PT

google_pt_mask = (data.google_lang=='pt')&(data.fasttext_lang!='pt')&(data.lang!='pt')
google_pt = data[google_pt_mask]
len(google_pt)

### Songs with only original attribute as PT

lang_pt_mask = (data.google_lang!='pt')&(data.fasttext_lang!='pt')&(data.lang=='pt')
lang_pt = data[lang_pt_mask]
len(lang_pt)

# Gathering PT sets

final_pt_data = pd.concat([three_way_pt, selected_lang_google_pt, 
                           selected_fasttext_pt]).drop(columns=["id", "fasttext_lang", "lang",
                                                                      "google_lang"])

final_pt_data.info()

final_pt_data.to_csv("data/MUSIC4ALL/processed/filtered_music4all_v3.csv", index=True)
