==> 4_unduplicated_crawling.ipynb <==
import pandas as pd
import time
import json

import spotipy

from spotipy.oauth2 import SpotifyClientCredentials

# Initial crawling

data = pd.read_csv("data/MUSIC4ALL/processed/filtered_music4all_v3.csv", index_col=0)
data.head()

# Duplicated detection by artist and title

data[data.duplicated(subset=["artist", "song"])]

# Duplicated detection by spotify_id

duplicated_id_mask = data.duplicated(keep=False, subset=["spotify_id"])
duplicated = data[duplicated_id_mask]
len(duplicated)

# Saving duplicated songs in one subset
duplicated.to_csv("data/MUSIC4ALL/processed/duplicated_ids.csv", index=True)

unique_data = data[~duplicated_id_mask]
len(unique_data)

# Saving unduplicated songs in another subset
unique_data.to_csv("data/MUSIC4ALL/processed/filtered_music4all_v4_part1.csv", index=True)

# Crawling Unique Data (Core)

unique_spotify_ids = unique_data.spotify_id.tolist()

max_tracks_per_call = 50
max_feats_per_call = 100
min_wait_time_per_call = 0.3

# Credentials should be added to system's path
credentials = SpotifyClientCredentials()

sp = spotipy.Spotify(auth_manager=credentials)

# Getting metadata for each unduplicated song
for start in range(0, len(unique_spotify_ids), max_tracks_per_call):
    end = start + max_tracks_per_call
    track_results = sp.tracks(unique_spotify_ids[start:end])
    
    if 'tracks' in track_results:  
        with open(f'data/MUSIC4ALL/crawled/track_results_{start}_{end-1}.json', 'w') as f:
            json.dump(track_results, f)
        print(f"Sucess")
    else:
        print(f"Error for indices {start} to {end-1}")
        
    time.sleep(min_wait_time_per_call)

# Getting audio features for each unduplicated song
for start in range(0, len(unique_spotify_ids), max_feats_per_call):
    end = start + max_feats_per_call
    audio_feat_results = sp.audio_features(unique_spotify_ids[start:end])
    
    if len(audio_feat_results) > 0:  
        with open(f'data/MUSIC4ALL/crawled/audio_feat_results_{start}_{end-1}.json', 'w') as f:
            json.dump(audio_feat_results, f)
        print(f"Sucess")
    else:
        print(f"Error for indices {start} to {end-1}")
        
    time.sleep(min_wait_time_per_call)
