==> 4_unduplicated_data_gathering.ipynb <==
import pandas as pd
import numpy as np

import time
import json
import glob
from tqdm import tqdm

from pprint import pprint

# Gathering of unduplicated songs attributes

data = pd.read_csv(f'data/MUSIC4ALL/processed/filtered_music4all_v4_part1.csv', index_col=0)
data.head()

# Reading audio feature jsons
feat_json = []

for json_path in tqdm(glob.glob("data/MUSIC4ALL/crawled/audio_feat_results*.json")):
    print(json_path)
    with open(json_path, 'r') as f:
        json_content = json.load(f)
        feat_json.extend(json_content)

len(feat_json)

len(data)

pprint(feat_json[0])

# Adding audio features to dataframe
FEAT_COLS = ['acousticness', 'danceability', 'duration_ms', 'energy', 'instrumentalness', 'key', 'liveness',
       'loudness', 'mode', 'speechiness', 'tempo', 'time_signature', 'valence']

for json_data in tqdm(feat_json):
    id_ = json_data['id']
    change_values = [json_data[key] for key in FEAT_COLS]
    data.loc[data.spotify_id == id_, FEAT_COLS] = change_values

data.head()

# Reading metadata jsons

info_json = []

for json_path in tqdm(glob.glob("data/MUSIC4ALL/crawled/track_results*.json")):
    print(json_path)
    with open(json_path, 'r') as f:
        json_content = json.load(f)['tracks']
        info_json.extend(json_content)

# Creating new metadata columns with null values
COLS_1 = ['album_name', 'song', 'artist', 'popularity', 'explicit', 'release', 'markets']
COLS_2 = [c +'_2' for c in COLS_1]

for col in COLS_2:
    data[col] = pd.NA

# Filling new columns
for cur_item in tqdm(info_json):    
    id_ = cur_item['id']
    
    album_name = cur_item['album']['name']
    song = cur_item['name']
    artist_name = cur_item['artists'][0]['name']
    popularity = cur_item['popularity']
    explicit = cur_item['explicit']
    markets = ','.join(cur_item['available_markets'])
    release = cur_item['album']['release_date']
    
    data.loc[data.spotify_id == id_, COLS_2] = [album_name, song, artist_name, popularity,
                                                explicit, release, markets]

# Updating metadata columns
for c in COLS_2:
    data[c[:-2]] = data[c]

final_core_data = data.drop(columns=COLS_2)

final_core_data.head()

final_core_data.to_csv('data/MUSIC4ALL/processed/filtered_music4all_v5_part1.csv', index=True)
