==> 5_music4all_duplicated_id_crawling.ipynb <==
import pandas as pd
import time
import json

import spotipy

from spotipy.oauth2 import SpotifyClientCredentials

import urllib.parse as uparse

from pprint import pprint

data = pd.read_csv("data/MUSIC4ALL/processed/duplicated_ids.csv", index_col=0)
data.head()

data[data.duplicated(keep=False, subset=['spotify_id'])].sort_values('spotify_id').head()

sorted_data_index = data.sort_index()

# Crawling IDs

min_wait_time_per_call = 0.5

credentials = SpotifyClientCredentials()

sp = spotipy.Spotify(auth_manager=credentials)

search_results = []

for index, row in sorted_data_index.iterrows():
    track = row['song']
    artist = row['artist']
    
    query = f"track:{track} artist:{artist}"
    
    try:
        qresult = sp.search(q=query, limit=1, type='track')
    except:
        qresult = {}
    
    if 'tracks' in qresult and 'items' in qresult['tracks']:
        print(f"Sucess for {index}")
    else:
        print(f"Error for track {track} and artist {artist}")
    
    search_results.append(qresult)
    time.sleep(min_wait_time_per_call)

len(search_results)

with open(f'data/MUSIC4ALL/crawled/raw_search_results.json', 'w') as f:
    json.dump(search_results, f)

processed_results = [sr['tracks']['items'] for sr in search_results]

with open(f'data/MUSIC4ALL/crawled/processed_search_results.json', 'w') as f:
    json.dump(processed_results, f)
