==> 6_duplicated_gathering_filtering.ipynb <==
import json

import pandas as pd

from pprint import pprint

# Deduplication of duplicated songs

duplicated = pd.read_csv(f'data/MUSIC4ALL/processed/duplicated_ids.csv', index_col=0)
duplicated = duplicated.sort_index()
duplicated.head()

COLS_1 = ['album_name', 'song', 'artist', 'popularity', 'spotify_id', 'explicit', 'markets', 'release']
COLS_2 = [c+'_2' for c in COLS_1]

for col in COLS_2:
    duplicated[col] = pd.NA

with open(f'data/MUSIC4ALL/crawled/raw_search_results.json', 'r') as f:
    content = json.load(f)

# Reading new metadata and adding to duplicated dataframe
for i, data in enumerate(content):
    cur_item = data['tracks']['items']
        
    if len(cur_item) == 1:
        cur_item = cur_item[0]
        
        album_name = cur_item['album']['name']
        song = cur_item['name']
        artist_name = cur_item['artists'][0]['name']
        popularity = cur_item['popularity']
        spotify_id = cur_item['id']
        explicit = cur_item['explicit']
        markets = cur_item['available_markets']
        release = cur_item['album']['release_date']
        duplicated.iloc[i,-8:] = [album_name, song, artist_name, popularity,
                                      spotify_id, explicit, markets, release]


duplicated.head()

# Remove empty ID cases

# Removing not found songs in spotify crawling
removed_empty_cases = duplicated[~duplicated.spotify_id_2.isnull()]
len(removed_empty_cases)

# Removed duplicated IDs and Lyrics

# Removing songs where both lyrics and spotify id are the same
removed_dup_lyrics = removed_empty_cases.drop_duplicates(subset=['lyrics', 'spotify_id_2'])
len(removed_dup_lyrics)

# Checking for inconsistencies
removed_dup_lyrics[removed_dup_lyrics.duplicated(keep=False, subset='spotify_id_2')].sort_values('spotify_id_2')

# There are still some inconsistencies, with duplicated spotify IDs and different lyrics
# To fix this, all remaining duplicated songs are removed
final_deduplicated = removed_dup_lyrics[~removed_dup_lyrics.duplicated(keep=False, subset='spotify_id_2')]
len(final_deduplicated)

# Saving deduplicated songs
final_deduplicated.to_csv(f'data/MUSIC4ALL/processed/deduplicated_ids.csv', index=True)
