==> 7_deduplicated_feats_crawling.ipynb <==
import pandas as pd

import time
import json

import spotipy

from spotipy.oauth2 import SpotifyClientCredentials

# Deduplicated metadata crawling

final_deduplicated = pd.read_csv(f'data/MUSIC4ALL/processed/deduplicated_ids.csv', index_col=0)
final_deduplicated.head()

len(final_deduplicated)

dedup_spotify_ids = final_deduplicated.spotify_id_2.tolist()

max_feats_per_call = 100
min_wait_time_per_call = 0.5

credentials = SpotifyClientCredentials()

sp = spotipy.Spotify(auth_manager=credentials)

# Crawling of deduplicated audio features
for start in range(0, len(dedup_spotify_ids), max_feats_per_call):
    end = start + max_feats_per_call
    audio_feat_results = sp.audio_features(dedup_spotify_ids[start:end])
    
    if len(audio_feat_results) > 0:  
        with open(f'data/MUSIC4ALL/crawled/dedup_feat_results_{start}_{end-1}.json', 'w') as f:
            json.dump(audio_feat_results, f)
        print(f"Sucess")
    else:
        print(f"Error for indices {start} to {end-1}")
        
    time.sleep(min_wait_time_per_call)
