==> 7_deduplicated_metadata_gathering.ipynb <==
import pandas as pd

import time
import json
import glob
from tqdm import tqdm

from pprint import pprint

# Deduplicated gathering

final_deduplicated = pd.read_csv(f'data/MUSIC4ALL/processed/deduplicated_ids.csv', index_col=0)
final_deduplicated.head()

# Reading audio features of deduplicated songs

all_json = []

for json_path in tqdm(glob.glob("data/MUSIC4ALL/crawled/dedup_feat_results*.json")):
    print(json_path)
    with open(json_path, 'r') as f:
        json_content = json.load(f)
        all_json.extend(json_content)

len(all_json)

pprint(all_json[0])

# Adding values to deduplicated dataframe
FEAT_COLS = ['acousticness', 'danceability', 'duration_ms', 'energy', 'instrumentalness', 'key', 'liveness',
       'loudness', 'mode', 'speechiness', 'tempo', 'time_signature', 'valence']

for i, data in enumerate(all_json):
    id_ = data['id']
    change_values = [data[key] for key in FEAT_COLS]
    final_deduplicated.loc[final_deduplicated.spotify_id_2 == id_, FEAT_COLS] = change_values

final_deduplicated.head()

final_deduplicated.columns

COLS_2 = [c for c in final_deduplicated.columns if c[-1]=='2']
COLS_2

# Updating columns
for c in COLS_2:
    final_deduplicated[c[:-2]] = final_deduplicated[c]

drop_final_deduplicated = final_deduplicated.drop(columns=COLS_2)

drop_final_deduplicated.head()

drop_final_deduplicated.columns

drop_final_deduplicated.to_csv(f'data/MUSIC4ALL/processed/deduplicated_ids_fixed_feat.csv', index=True)
