==> 8_full_data_gathering.ipynb <==
import pandas as pd

# Unduplicated (core) and deduplicated (complement) set gathering

core = pd.read_csv("data/MUSIC4ALL/processed/filtered_music4all_v5_part1.csv", index_col=0)
core.info()

complement = pd.read_csv("data/MUSIC4ALL/processed/deduplicated_ids_fixed_feat.csv", index_col=0)
complement.info()

all(core.columns == complement.columns)

# Core Treatment

core.head()

core.time_signature = core.time_signature.astype(int)

core.explicit = core.explicit.astype(int)

core.genres = core.genres.apply(lambda s: s.strip().lower().split(','))

core["nb_genres"] = core.genres.apply(lambda l: len(l))

core['year'] = core.release.apply(lambda s: s[:4])

core.year = core.year.astype(int)

core[['key', 'mode', 'valence', 'tempo']].head()

core.key = core.key.astype(int)

core['mode'] = core['mode'].astype(int)

core.head()

core[['key', 'mode', 'valence', 'tempo']].head()

final_core = core.drop(columns="release")

final_core.info()

# Complement Treatment

complement.head()

complement.time_signature = complement.time_signature.astype(int)

complement.explicit = complement.explicit.astype(int)

complement.genres = complement.genres.apply(lambda s: s.strip().lower().split(','))

complement["nb_genres"] = complement.genres.apply(lambda l: len(l))

complement['year'] = complement.release.apply(lambda s: s[:4])

complement.year = complement.year.astype(int)

complement.markets = complement.markets.apply(lambda s: s.strip()\
                                                         .replace('[', '')\
                                                         .replace(']', '')\
                                                         .replace(' ', '')\
                                                         .replace('\'', ''))

complement[['key', 'mode', 'valence', 'tempo']].head()

complement.key = complement.key.astype(int)

complement['mode'] = complement['mode'].astype(int)

complement.head()

complement[['key', 'mode', 'valence', 'tempo']].head()

final_complement = complement.drop(columns="release")

final_complement.info()

# Merge

final_music4all = pd.concat([final_core, final_complement])

final_music4all = final_music4all.sort_index()

final_music4all.markets = final_music4all.markets.fillna("UNK")

final_music4all.info()

final_music4all.to_csv("data/MUSIC4ALL/processed/treated_music4all_v1.csv", index=True)
