==> 92_full_data_duplicated_lyrics_filtering.ipynb <==
import pandas as pd
import spacy

import json

from pprint import pprint

from tqdm.notebook import tqdm

tqdm.pandas()

# Filtering duplicated lyrics after merging subsets

data = pd.read_csv("data/MUSIC4ALL/processed/treated_music4all_v2.csv", index_col=0)

data.info()

# Spotify ID Analysis

data[data.duplicated(keep=False, subset=['spotify_id'])].sort_values("spotify_id")

data = data.drop(labels=[38040, 54904])

data[data.duplicated(keep=False, subset=['spotify_id'])].sort_values("spotify_id")

# Exact Lyrics Analysis

data[data.duplicated(keep=False, subset=['lyrics'])].sort_values("lyrics")[['lyrics', 'artist', 'song', 'popularity', 'year']]

# Manually selected rows to drop (always the newest to be released)
data = data.drop(labels=[44076, 74874, 41950, 4051, 45250, 103652, 61877, 52688, 69416, 94667, 86460])

data[data.duplicated(keep=False, subset=['lyrics'])].sort_values("lyrics")[['lyrics', 'artist', 'song', 'popularity', 'year']]

len(data)

# Approximate Lyric Similarity

lyrics_df = data.copy()[['lyrics']]

#!python -m spacy download pt_core_news_md

nlp = spacy.load('pt_core_news_md')

lyrics_df["lyrics_obj"] = lyrics_df.lyrics.str.lower().progress_apply(nlp)

matches = {}

# Create graph alike structure based on cosine similarity between lyrics

for i, row_i in enumerate(tqdm(lyrics_df.iterrows())):
    index_i, data_i = row_i

    for index_j, data_j in lyrics_df.iloc[i+1:].iterrows():
        
        score = data_i.lyrics_obj.similarity(data_j.lyrics_obj)
        
        if score >= 0.99:
            if index_i not in matches:
                matches[index_i] = {index_j: score}
            else:
                matches[index_i][index_j] = score

            if index_j not in matches:
                matches[index_j] = {index_i: score}
            else:
                matches[index_j][index_i] = score

with open('data/MUSIC4ALL/processed/lyrics_matches.json', 'w') as f:
    json.dump(matches, f)

### Removing

with open('data/MUSIC4ALL/processed/lyrics_matches.json', 'r') as f:
    matches = json.load(f)

len(matches.keys())

# Function to choose priority song based n a list of similar ones
def choose_priority_item(idxs):
        
    if len(idxs) == 1:
        return idxs[0]
        
    chosen_idx = None
    i = 0
    j = 1

    while i < len(idxs) and j < len(idxs):

        if i == j:
            j+=1
            continue
                
        if data[data.index==int(idxs[i])].year.values[0] <= data[data.index==int(idxs[j])].year.values[0]:
            chosen_idx = idxs[i]
            j+=1
        else:
            chosen_idx = idxs[j]
            i+=1
    
    return chosen_idx

remove_idxs = set()

# Finding songs to remove
for key, values in tqdm(matches.items()):
    complement = [k for k in values.keys() if k not in remove_idxs]
    
    if key not in remove_idxs:
        possibles = [key, *complement]
    else:
        possibles = [*complement]
    
    if len(possibles) > 0:
        chosen_idx = choose_priority_item(possibles)
        possibles.remove(chosen_idx)
        remove_idxs.update(possibles)

len(remove_idxs)

int_idxs = [int(idx) for idx in remove_idxs]

final_data = data.drop(labels=int_idxs)

len(final_data)

final_data.columns

final_data.to_csv("data/MUSIC4ALL/processed/treated_music4all_v3.csv", index=True)
