==> 93_tcc_dataset.ipynb <==
import pandas as pd
import spacy

from ast import literal_eval
from collections import Counter
from tqdm import tqdm

tqdm.pandas()

import json

from pprint import pprint

# Finalizing dataset

data = pd.read_csv("data/MUSIC4ALL/processed/treated_music4all_v3.csv", index_col=0)

data['genres'] = data.genres.progress_apply(literal_eval)

data.info()

data = data.drop(columns=["markets"])

# Counting genres
genres = []

for i,d in data.iterrows():
    genres.extend(d.genres)

genre_count = Counter(genres)

chosen_genre = set([t[0] for t in genre_count.most_common(20)])
chosen_genre

# Checking if all all songs have at least one of the top 20 genres
true = 0

for i, d in data.iterrows():
    if len(set(d.genres).intersection(chosen_genre)) > 0:
        true+=1
        
print(true==len(data))

data.to_csv("data/MUSIC4ALL/processed/tcc_music4all.csv", index=True)
